+++ lib/abstract_mapper/rspec.rb
@@ -3,5 +3,4 @@
 # Collection of shared examples for testing parts of DSL
 require_relative "../rspec/nodes"
 require_relative "../rspec/rules"
-require_relative "../rspec/mapper"
 require_relative "../rspec/doubles"